third_party/WebKit/LayoutTests/inspector/view-events-expected.txt

This tests that events are properly propagated through View hierarchy.


Running: testShowView
View()
View.show()
  View.wasShown()
  View.onResize()
  View.cacheSize()
View.detach()
  View.discardCachedSize()
  View.willHide()

Running: testAppendViaDOM
View()
Error: Attempt to add view via regular DOM operation.

Running: testInsertViaDOM
View()
Error: Attempt to add view via regular DOM operation.

Running: testAttachToOrphanNode
View()
View.show()
Error: Attempt to attach view to orphan node

Running: testImmediateParent
Parent()
Child()
Child.show()
OK

Running: testDistantParent
Parent()
Child()
Child.show()
OK

Running: testEvents
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Parent.doResize()
  Parent.discardCachedSize()
  Parent.cacheSize()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Parent.doResize()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Child.wasShown()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Child.detach()
  Child.discardCachedSize()
  Child.willHide()
Parent.detach()
  Parent.discardCachedSize()
  Parent.willHide()

Running: testEventsHideOnDetach
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Parent.doResize()
  Parent.discardCachedSize()
  Parent.cacheSize()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Parent.doResize()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Child.wasShown()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Child.detach()
  Child.discardCachedSize()
  Child.willHide()
Parent.detach()
  Parent.discardCachedSize()
  Parent.willHide()

Running: testViewCounter
Parent()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Child()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
  view counter: 1
Child 2()
Child 2.show()
  Child 2.wasShown()
  Child 2.onResize()
  Child 2.cacheSize()
  view counter: 2
Child.detach()
  Child.discardCachedSize()
  Child.willHide()
  view counter: 1
Child 2.detach()
  Child 2.discardCachedSize()
  Child 2.willHide()
  view counter: 0

Running: testRemoveChild
Parent()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Child()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Error: Attempt to remove element containing view via regular DOM operation

Running: testImplicitRemoveChild
Parent()
Child()
Child.show()
Error: Attempt to remove element containing view via regular DOM operation

Running: testRemoveChildren
Parent()
Child()
Child.show()
Error: Attempt to remove element containing view via regular DOM operation

Running: testImplicitRemoveChildren
Parent()
Child()
Child.show()
Error: Attempt to remove element containing view via regular DOM operation

Running: testShowOnWasShown
Parent()
Child()
Parent.show()
  Parent.wasShown()
Child.show()
  Child.cacheSize()
  Parent.onResize()
  Child.wasShown()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()

Running: testShowNestedOnWasShown
Top()
Middle()
Bottom()
Middle.show()
Top.show()
  Top.wasShown()
Bottom.show()
  Bottom.cacheSize()
  Top.onResize()
  Middle.wasShown()
  Middle.onResize()
  Bottom.wasShown()
  Bottom.onResize()
  Top.cacheSize()
  Middle.cacheSize()
  Bottom.cacheSize()
Top.detach()
  Top.discardCachedSize()
  Middle.discardCachedSize()
  Bottom.discardCachedSize()
  Bottom.willHide()
  Middle.willHide()
  Top.willHide()

Running: testDetachOnWasShown
Parent()
Child()
Child.show()
Parent.show()
  Parent.wasShown()
Child.detach()
  Child.discardCachedSize()
  Parent.onResize()
  Parent.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Parent.willHide()

Running: testShowOnWillHide
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()
Child.show()

Running: testDetachOnWillHide
Parent()
Child()
Parent.show()
  Parent.wasShown()
  Parent.onResize()
  Parent.cacheSize()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()
Child.detach()
  Child.discardCachedSize()

Running: testShowDetachesFromPrevious
Parent1()
Parent2()
Child()
Parent1.show()
  Parent1.wasShown()
  Parent1.onResize()
  Parent1.cacheSize()
Parent2.show()
  Parent2.wasShown()
  Parent2.onResize()
  Parent2.cacheSize()
Child.show()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()
Child.show()
Child.detach()
  Child.discardCachedSize()
  Child.willHide()
  Child.wasShown()
  Child.onResize()
  Child.cacheSize()

Running: testResizeOnWasShown
Parent()
Child()
Child.show()
Parent.show()
  Parent.wasShown()
Child.doResize()
  Child.discardCachedSize()
  Child.cacheSize()
  Parent.onResize()
  Child.wasShown()
  Child.onResize()
  Parent.cacheSize()
  Child.cacheSize()
Parent.detach()
  Parent.discardCachedSize()
  Child.discardCachedSize()
  Child.willHide()
  Parent.willHide()
